+++ lib/constants.dart
@@ -18,7 +18,8 @@ String _androidMainActivityKitFile(
         String domain, String company, String appname) =>
     'android/app/src/main/kotlin/$domain/$company/$appname/MainActivity.kt';
 String applicationName = '\${applicationName}';
-String introMessage(String currentVersion) => '''
+String introMessage(String currentVersion) =>
+    '''
   ════════════════════════════════════════════
      ANIMATED NATIVE SPLASH (v$currentVersion)
   ════════════════════════════════════════════
@@ -29,4 +30,6 @@ const String _webFolder = 'web/';
 const String _webSplashFolder = _webFolder + 'splash/';
 const String _webSplashImagesFolder = _webSplashFolder + 'img/';
 const String _webIndex = _webFolder + 'index.html';
-const String _webRelativeStyleFile = 'splash/style.css';
+const String _webRelativeStyleFile = 'web/splash/style.css';
+const String _url = '\$FLUTTER_BASE_HREF';
+String _projectName = '';